+++ src/app/macros_config.cpp
@@ -20,7 +20,7 @@
 #define KEY_BLOB  "b"
 
 #define MACROS_MAGIC 0x4D414352u // 'MACR'
-#define MACROS_VERSION 7
+#define MACROS_VERSION 8
 
 static Preferences prefs;
 
@@ -185,7 +185,7 @@ void macros_config_set_defaults(MacroConfig* cfg) {
         for (int b = 0; b < MACROS_BUTTONS_PER_SCREEN; b++) {
             cfg->buttons[s][b].action = MacroButtonAction::None;
             cfg->buttons[s][b].label[0] = '\0';
-            cfg->buttons[s][b].script[0] = '\0';
+            cfg->buttons[s][b].payload[0] = '\0';
 
             cfg->buttons[s][b].icon.type = MacroIconType::None;
             cfg->buttons[s][b].icon.id[0] = '\0';